Changed UnitNumbersService to UnitNumberService, its path from unit-numbers to unit-number, getAll to list, getOne to retrieve; Updated imports to new path and class; Added docstrings to UnitNumber related services, model;
