+++ web/src/app/components/units/unit-add/unit-add.component.ts
@@ -7,7 +7,7 @@ import { NotifyService } from 'src/app/shared/handlers/notify/notify.service';
 
 import { Block } from 'src/app/shared/services/block/block.model';
 import { Floor } from 'src/app/shared/services/floor/floor.model';
-import { UnitNumber } from 'src/app/shared/services/unit-numbers/unit-numbers.model';
+import { UnitNumber } from 'src/app/shared/services/unit-number/unit-number.model';
 import { UnitsService } from 'src/app/shared/services/units/units.service';
 
 @Component({
