+++ web/src/app/shared/services/unit-number/unit-number.service.ts
@@ -3,14 +3,17 @@ import { environment } from 'src/environments/environment';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
-import { UnitNumber } from './unit-numbers.model';
+import { UnitNumber } from './unit-number.model';
 
 const BASE_URL = `${ environment.baseUrl }v1/unit-numbers/`
 
+/**
+ * A service for UnitNumber related methods
+ */
 @Injectable({
   providedIn: 'root'
 })
-export class UnitNumbersService {
+export class UnitNumberService {
 
   // Data
   public unitNumber: UnitNumber | undefined
@@ -20,6 +23,14 @@ export class UnitNumbersService {
     private http: HttpClient
   ) { }
 
+  /**
+   * Create new unit number entry
+   *
+   * @param body - payload
+   * @param body.unitNumber - unit number name
+   * 
+   * @returns New created unit number entry
+   */
   create(body: any): Observable<UnitNumber> {
     const urlTemp = `${ BASE_URL }`
     return this.http.post<UnitNumber>(urlTemp, body).pipe(
@@ -30,7 +41,12 @@ export class UnitNumbersService {
     )
   }
 
-  getAll(): Observable<UnitNumber[]> {
+  /**
+   * Returns a list of unit numbers
+   *
+   * @returns List of unit numbers
+   */
+  list(): Observable<UnitNumber[]> {
     const urlTemp = `${ BASE_URL }`
     return this.http.get<UnitNumber[]>(urlTemp).pipe(
       tap((res: UnitNumber[]) => {
@@ -40,7 +56,14 @@ export class UnitNumbersService {
     )
   }
 
-  getOne(id: string | undefined): Observable<UnitNumber> {
+  /**
+   * Returns a unit number
+   * 
+   * @param id - unit number ID
+   *
+   * @returns A unit number
+   */
+  retrieve(id: string | undefined): Observable<UnitNumber> {
     const urlTemp = `${ BASE_URL }${ id }/`
     return this.http.get<UnitNumber>(urlTemp).pipe(
       tap((res: UnitNumber) => {
@@ -50,6 +73,15 @@ export class UnitNumbersService {
     )
   }
 
+  /**
+   * Partial update a unit number
+   *
+   * @param id - unit number ID
+   * @param body - payload
+   * @param body.unitNumber - unit number name
+   * 
+   * @returns Updated unit number
+   */
   patch(id: number, body: any): Observable<UnitNumber> {
     const urlTemp = `${ BASE_URL }${ id }/`
     return this.http.patch<UnitNumber>(urlTemp, body).pipe(
