--- web/src/app/shared/services/unit-numbers/unit-numbers.model.ts
@@ -1,11 +1,0 @@
-export class UnitNumber {
-    constructor(
-        public id: number,
-        public unitNumber: string,
-        public isActive: boolean,
-        public createdAt: string,
-        public createdBy: number,
-        public lastModifiedAt: string,
-        public lastModifiedBy: number
-    ) {}
-}
